chore: Update dependencies
The motivation was that the ancient version of the graphql library was
emitting a bogus, old syntax -- "implements A, B" instead of "A & B".
This change just runs "yarn upgrade" and makes the minimal changes to
get the result to pass the type checker and unit tests. I know less
than nothing about the NodeJS extended universe, so shout-out to LLMs
for letting me stay mostly ignorant.

Of course, there are still dozens of vulnerable or deprecated
dependencies, but whatever.